+++ src/engine/kinematics/index.ts
@@ -3,3 +3,4 @@ export * from './emaFilter';
 export * from './parallaxCheck';
 export * from './poseDetector';
 export * from './occlusionHeuristics';
+export * from './rulesEngine';